02/requirements.txt
pillow-9.2.0